feat(init): boilerplate for checking exercises
This is a really basic approach to code exercises, but it should
do for this workshop. Effectively, each exercise is a refactoring
challenge with an expected output object. The result of the
refactor is checked against the expected solution (after running
both through `JSON.stringify`) to verify if it’s correct or not,
then both are displayed on the page with a visual indication of
whether or not the solution is correct.